use jitpack to avoid need for local builds of test suites
I tried integrating the jitpack.io dependencies for the test suites into
the Maven build, which in principle worked but somehow Maven asked for
additional dependencies which I could not solve in Maven.

Thus the build was changed to Gradle, and the `production` and `beta`
variants are now configured solely through the test suite names and
versions.